SB-536: Implement cron tasks
- Class moved
