+++ strongbox-cron-tasks-core/src/main/java/org/carlspring/strongbox/cron/api/jobs/JavaCronJob.java
@@ -1,0 +1,9 @@
+package org.carlspring.strongbox.cron.api.jobs;
+
+/**
+ * @author carlspring
+ */
+public abstract class JavaCronJob extends AbstractCronJob
+{
+
+}
