+++ strongbox-cron-tasks-core/src/main/java/org/carlspring/strongbox/cron/exceptions/CronTaskException.java
@@ -1,0 +1,30 @@
+package org.carlspring.strongbox.cron.exceptions;
+
+/**
+ * @author Yougeshwar
+ */
+public class CronTaskException
+        extends Exception
+{
+
+    public CronTaskException()
+    {
+    }
+
+    public CronTaskException(String msg)
+    {
+        super(msg);
+    }
+
+    public CronTaskException(String msg,
+                             Throwable cause)
+    {
+        super(msg, cause);
+    }
+
+    public CronTaskException(Throwable cause)
+    {
+        super(cause);
+    }
+
+}
