+++ strongbox-cron-tasks-core/src/main/java/org/carlspring/strongbox/cron/exceptions/CronTaskNotFoundException.java
@@ -1,0 +1,30 @@
+package org.carlspring.strongbox.cron.exceptions;
+
+/**
+ * @author Yougeshwar
+ */
+public class CronTaskNotFoundException
+        extends Exception
+{
+
+    public CronTaskNotFoundException()
+    {
+    }
+
+    public CronTaskNotFoundException(String msg)
+    {
+        super(msg);
+    }
+
+    public CronTaskNotFoundException(String msg,
+                                     Throwable cause)
+    {
+        super(msg, cause);
+    }
+
+    public CronTaskNotFoundException(Throwable cause)
+    {
+        super(cause);
+    }
+
+}
