--- strongbox-cron-tasks-core/src/main/java/org/carlspring/strongbox/crontask/api/jobs/JavaCronJob.java
@@ -1,9 +1,0 @@
-package org.carlspring.strongbox.crontask.api.jobs;
-
-/**
- * @author carlspring
- */
-public abstract class JavaCronJob extends AbstractCronJob
-{
-
-}
