--- strongbox-cron-tasks-core/src/main/java/org/carlspring/strongbox/crontask/exceptions/CronTaskException.java
@@ -1,30 +1,0 @@
-package org.carlspring.strongbox.crontask.exceptions;
-
-/**
- * @author Yougeshwar
- */
-public class CronTaskException
-        extends Exception
-{
-
-    public CronTaskException()
-    {
-    }
-
-    public CronTaskException(String msg)
-    {
-        super(msg);
-    }
-
-    public CronTaskException(String msg,
-                             Throwable cause)
-    {
-        super(msg, cause);
-    }
-
-    public CronTaskException(Throwable cause)
-    {
-        super(cause);
-    }
-
-}
