--- strongbox-cron-tasks-core/src/main/java/org/carlspring/strongbox/crontask/exceptions/CronTaskNotFoundException.java
@@ -1,30 +1,0 @@
-package org.carlspring.strongbox.crontask.exceptions;
-
-/**
- * @author Yougeshwar
- */
-public class CronTaskNotFoundException
-        extends Exception
-{
-
-    public CronTaskNotFoundException()
-    {
-    }
-
-    public CronTaskNotFoundException(String msg)
-    {
-        super(msg);
-    }
-
-    public CronTaskNotFoundException(String msg,
-                                     Throwable cause)
-    {
-        super(msg, cause);
-    }
-
-    public CronTaskNotFoundException(Throwable cause)
-    {
-        super(cause);
-    }
-
-}
