--- strongbox-cron-tasks-core/src/main/java/org/carlspring/strongbox/crontask/repository/CronTaskConfigurationRepository.java
@@ -1,17 +1,0 @@
-package org.carlspring.strongbox.crontask.repository;
-
-import org.carlspring.strongbox.crontask.domain.CronTaskConfiguration;
-import org.carlspring.strongbox.data.repository.OrientRepository;
-
-import org.springframework.transaction.annotation.Transactional;
-
-/**
- * @author Yougeshwar
- */
-@Transactional
-public interface CronTaskConfigurationRepository
-        extends OrientRepository<CronTaskConfiguration>
-{
-
-    CronTaskConfiguration findByName(String name);
-}
