--- strongbox-cron-tasks-core/src/main/java/org/carlspring/strongbox/crontask/services/CronTaskDataService.java
@@ -1,19 +1,0 @@
-package org.carlspring.strongbox.crontask.services;
-
-import org.carlspring.strongbox.crontask.domain.CronTaskConfiguration;
-import org.carlspring.strongbox.data.service.CrudService;
-
-import java.util.Optional;
-
-import org.springframework.transaction.annotation.Transactional;
-
-/**
- * @author Yougeshwar
- */
-@Transactional
-public interface CronTaskDataService
-        extends CrudService<CronTaskConfiguration, String>
-{
-
-    Optional<CronTaskConfiguration> findByName(final String name);
-}
